build and help commands mostly completed
creates the help command and tweaks Io so its more
useful for printing to the terminal.

cleans up build and limits its functionality to
checking for version mismatches, attempting to easily
rectify it, but if it can't then it will fail.

need to create some more commands to allow for
making a lock file and maintaining it.