refactor: creates command runners for apps
made a `runCommand` function for both haxe and
haxelib so that it will handle the running and
return a result of success,error and the raw app
output.

made a new class called `Result` that will be used
to pass the outcome of the command plus the output
if we want it.

and made a new switch called `-o|--output` that will
cause this to be written to the screen when used
so we can troubleshoot issues if things aren't working
as expected.