fix: changes default output of Haxelib from error to ok
if haxelib runs a command and everything is great
sometimes it doesn't output anything because there
is nothing to say. in these cases i was automatically
labelling it as an error because i was assuming that
it had to output something. so i changed it to
return an `Ok` in this case.