feat: adds upgrade
adds upgrade functionality that will upgrade a
library to a newer version and check if the project
still builds.

if the project builds successfully then that new
version is set as the current locked version.

if that project fails then the last version that
allowed the project to build is used.

if the user wants to push to a new failing version
they will have to use the `set` command as it doesn't
check build status.