fix: lockfile malform check, and doesn't make lockfile if no arguments
fixes an issue when running `haxelock lock` without any arguments it
will create a blank haxe.lock file & show the help. now it just shows
help.

fixes an issue where it crashes when a haxe.lock file exists but isn't
formatted as expected. now it throws an error saying that the file is
malformed and that the user should remove and recreate the file. this
shouldn't be a production issue, but currently occurs because the format
for the haxe.lock file is in flux.