refactor: upgrade will only test when actually upgrading
changes the output of the upgradeLibrary function to an enum which details
if the upgrade was sucessful, if there was no need for upgrade, and if
the upgrade failed. this way it now has enough information to check if
there was actually an upgrade before building and ensuring that the builds
all pass.